Keep all inbox notifications related types in one module
