+++ packages/liveblocks-core/src/store.ts
@@ -5,8 +5,8 @@ import type { Resolve } from "./lib/Resolve";
 import type { BaseMetadata } from "./protocol/Comments";
 import type { CommentData, CommentReaction } from "./protocol/Comments";
 import type { CommentUserReaction } from "./protocol/Comments";
-import type { InboxNotificationData } from "./types/InboxNotificationData";
-import type { InboxNotificationDeleteInfo } from "./types/InboxNotificationDeleteInfo";
+import type { InboxNotificationData } from "./protocol/InboxNotifications";
+import type { InboxNotificationDeleteInfo } from "./protocol/InboxNotifications";
 import type { PartialNullable } from "./types/PartialNullable";
 import type { RoomNotificationSettings } from "./types/RoomNotificationSettings";
 import type { ThreadData, ThreadDataWithDeleteInfo } from "./protocol/Comments";
