--- packages/liveblocks-core/src/types/InboxNotificationDeleteInfo.ts
@@ -1,11 +1,0 @@
-import type { DateToString } from "../lib/DateToString";
-
-export type InboxNotificationDeleteInfo = {
-  type: "deletedInboxNotification";
-  id: string;
-  roomId: string;
-  deletedAt: Date;
-};
-
-export type InboxNotificationDeleteInfoPlain =
-  DateToString<InboxNotificationDeleteInfo>;
